chore: Reformat the code of Playground
